+++ Ogre3DSeparateBuildSpace/VinezorsiOS/VinezorsiOS/EngineStage.h
@@ -29,9 +29,9 @@ public:
     }
     
     virtual void activatePerformLeftMove();
-    virtual void activatePerformLeftMove(int angle);
+    virtual void activatePerformLeftMove(float angle);
     virtual void activatePerformRightMove();
-    virtual void activatePerformRightMove(int angle);
+    virtual void activatePerformRightMove(float angle);
     virtual void activatePerformSwipeUp();
     virtual void activatePerformSwipeDown();
     virtual void activatePerformDoubleTap(float x, float y);
